delete cart

+++ order/src/main/java/com/chill/order/service/CartService.java
@@ -34,7 +34,20 @@ public class CartService {
     }
     @CacheEvict(value = "cart_cache", key = "#cartId")
     public void deleteCart(int cartId) {
-        cartRepository.deleteById(cartId);
+        Optional<Cart> cart= cartRepository.findById(cartId);
+        if (cart.isPresent()) {
+            List<MysteryBagDTO> mysteryBags= cart.get().getProducts();
+            for(MysteryBagDTO mysteryBag:mysteryBags) {
+                mysteryBagClient.getMysteryBag(mysteryBag.getId(),(-1));
+            }
+            cartRepository.deleteById(cartId);
+        }
+        else{
+            throw new RuntimeException("Cart not found");
+        }
+
+
+
     }
 
     @Cacheable(value = "cart_cache", key = "#cartId")
